fix: selectRow returns false
git-svn-id: http://halodev.ontoprise.com/svn/HaloSMWExtension/trunk@13049 f80b7aa0-dfcc-4e65-a0d9-74108174cbc8

+++ extensions/SMWHalo/smwtsc/includes/triplestore_client/TSC_HaloStore2.php
@@ -73,7 +73,7 @@ class SMWHaloStore2 extends SMWStoreAdapter {
 			$id = $db->selectRow($smw_ids, array('smw_id'), array('smw_title'=>$subjectTitle->getDBkey(), 'smw_namespace'=>$subjectTitle->getNamespace()));
 		}
 
-		if (is_null($id)) return; // something is wrong. stop here
+		if (is_null($id) || $id === false) return; // something is wrong. stop here
 
 		// delete old mappings
 		$db->delete($smw_urimapping, array('smw_id' => $id->smw_id));
